pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-integration-agent"
version = "1.0.0"
description = "AI agent for integrating with Jira, Confluence, and Java code processing"
requires-python = ">=3.12"
authors = [
    {name = "AI Agent Developer", email = "developer@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests>=2.31.0",
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "javalang>=0.13.0",
    "atlassian-python-api>=3.41.0",
    "pyyaml>=6.0.1",
    "click>=8.1.7",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "black>=23.0", "flake8>=6.0"]

[project.scripts]
ai-agent = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"